requirements.txt
python-vlc
ffmpeg-python